backend/scripts/requirements.txt
psycopg2-binary>=2.9.0
faker>=30.0.0
